feat(transaction): refetch vault balance after send or decline transaction
