+++ src/modules/vault/hooks/assets/index.ts
@@ -1,2 +1,3 @@
+export * from './useHasReservedCoins';
 export * from './useVaultAssets';
 export * from './useVaultAssetsList';
